Initial commit of Matlab simulation
Working versions of matlab simulations of the critically sampled and oversampled
polyphase filterbank are working.

The version in bjf/ was provided by Dr. Jeffs. The version in mat/ is an
independent solution. They have not been compared to each other but individually
it looks like the results are correct.

The mat/ version needs to be cleaned up and still have some documentation notes
to add. But this is a working version that can now lead to c code development.