Start of C code
Briefly started to add a C/C++ version of the MATLAB code. The approach was to try
and translate as much as possible to C/C++. However, decided to go ahead and
start to combine HLS at the same time since the Vivado HLS guide shows examples
for how to use the built in FFT IP.

The idea will be to come back and get a benchmarked C version as well.