MATLAB read C sim data
Modified the sim file to be flexible to an arbitrary data type. This way HLS can
start to be used a long with other standard C types. The file writing mechanism
was adjusted to write in binary instead of ASCII and to support the data_t.

Added a matlab script to parse the binary file and plot results for comparison.