Update prototype LP filter representation
Learned that C++ does not support (in the standard) of interpreting hex values
as floats. They are convered to unsigned long ints. Also removed its definition
as a header file and insted as data file in the data directory. They might come
back as a header to store long term, just not for now.

Since hex floats are not supported the matlab script was modified to write
decimal floats and instead include the coeff.dat file in the implementation
code.