Working bit level accurate simulation
Next step is to attempt synthesis and fix any problem that come up from that.

These changes submit what seemms to be a working version of a C implementation
of the Oversampled PFB.

The primary contribution here is the addition of the Vivado HLS FFT IP in the C
code. This includes the function call, setting up the configuration (bit width,
FFT size, bit ordering), status parameters, transform direction (e.g., inverse
for the os pfb). This also includes an overflow signal.

Some questions are still left open:
  * Is there any way to template a static array for initializaion at compile
    so that the shift array can have values loaded in it already instead of
    passed in as a parameter?
  * Right now the operating assumption is to use the model of a shift register
    for both the input data and the frequenc compensation buffer. I have not
    really considered other methods to satisfy these. Using a shift register
    seems to make sense for the input data but the frequncy shift maybe could
    just be a direct memory access. This is because we are actually shifting one
    element at a time. E.g., for a shift of 8 we use a for loop to shift one up
    one element 8 times instead of just making the one time copy. In my mind
    this is an obvious optimization. In order to do this I would need to have
    second buffer to copy to.
  * Should consider refactoring the code to mimic the FFT single example that
    uses dummy_fe and dummy_be functions to seperate the dataflow. This could
    help when doing things like the frequncy correction step.
  * Need to look into doing fixed point instead of floating
  * Need to run the input data through my Matlab code to produce sort of an
    expected output (however, I think matlab is double precision and so I will
    need to learn how to force it to single precision to try hopefully I get an
    exact match between the C version and matlab).

The simulation/testbench file also opens a data file to dump the results for
post processing/plotting with an associated matlab script.