Branch starting streaming data ordering
Modified approach to order data in arrays using a streaming mentatlity which is
that the earliest data is at the end of an array as opposed to the natural
ordering where the first data point is the first element in the data container.
This is done to better format and model the code matching a streaming
architecture in the hardware design.

This is done on a new branch because I also want to maintain the natural
ordering version as a comparison because I learned a lot from it and I believe
it is a good comparison. The main changes to make the streaming ordering work is
also documented in the Matlab simulation and to achieve the same result in C
required the same steps.
  * The input data needs to be flipped - this is achieved in C by just filling
    the array from the end to the front and then decrementing a data pointer
    instead of advancing one
  * The shift register (matlab circshift) needs to shift up and out instead of
    down. This is again achieved by starting at the end of the array and copying
    up the array. Something like this is commonly seen in FIR and shift register
    examples for HLS.
  * The multiplication of the filter_state var and filter coefficeints need to
  * now be indexed the same.