Working version that synthesizes
This commit is a milestone that produces a result that Vivado HLS can synthesize
to hardware. It continues off the previous work to satisfy a streaming
mentality. The work to get here was helped by really thinking about the FIR
examples and keeping in mind that the OS PFB is a decimating filter. I knew that
Vivado offers pipelined decimation filters and that there had to be a way to get
my implementation and the algorithm to conform.

The big difference now is that instead of having 3 loops that are responsible
for shifting the filter state, copying in new data, and doing the multiply
accumulate, respectively that we now bring those steps inside one loop and
operate on each sample individually. However, in order to get this to work
multiply accumulates need to have memory accesses that are lineary (i.e., match
the streaming model of one sample in increasing order). And it took me a while
to see how to do that but the trick was to look at the indexing scheme for the
filter state and coefficient variables. The expression shows that p*M+m is the
way to map sampels to their location and since m is the fastest moving element
the way to achieve a linear memory access is to interchange the P and M loops.
Meaning that m is now the inner loop with P being the outer loop. This creates a
linear memory access as if the bounds on the for loop where (int i=L-1; i>=0;
++i) like we had to do previously to match streaming paradigm.

This commit also has other changes that were supposed to help me get to test the
iplementation on hardware. For example, the shift_states is now hardcoded in
because I didn't want to risk the update to the array to break the dataflow
assumption (according to a synthesis error). Then I wanted to use axi streams
(because at the time I thought it was the only way to interface with the CPU)
and this required the addition of an out variable that since the hls::fft
implements an ap_fifo interface instead.

I was also reminded that the c++ struct inheritance didn't default inherit the
logic to implement all the required configuration settings (due to the lack of a
constructor). And so you have to implement the stages_block_ram member if the
max_nfft value is changed.