Implement AXIS side channels for DMA access
The OS PFB has a streaming pipelined implementation and in order to poke at and
probe if the core is functional I have chosen to use PYNQ, but this is running
on the ARM processor and is a memory mapped region. Therefore to talk to the
core and use the processor we need a DMA and to use the DMA the AXIS protocol
needs the TLAST handshake signal. Without the TLAST signal the DMA hangs and
does not finish the transaction.