Modify to use float on the input becuase of FFT core (-1, 1) scale requirement
The Vivado FFT block states that the input values should be betwenn the range of
(-1, 1). I implemented this in a practice FFT project and trasnferred it here.
However our data products on instrument will be 8-bit real/8-bit imaginary and
so I need to look into how I can do this. I am positive I will have to rely on
the ap_fixed point library since that is the only way I know how to have a 8-bit
number with a fixed point at 7.

The quantization approach implemented here results in a bias that appears in the
0 Hz bin and I think this because of the subtract by 0.5 (and this may actually
be commonly referred to in our group by the half-LSB bias). This needs to be
looked into and understood.