Remove decimate by 16 and comment on dataflow changes
Silly things happen when you don't commit your code after awhile. You forget
what is going on.

So the commit prior to this one talks about a decimate by 16 version to look at
DMA bugs but what it forgot to mention was that it had significant code changes
to restructure the function calls to better match the dataflow optimization
requirement by HLS in synthesis.

The requirement is that there is a one producer consumer relationship and this
is what I can't remember what I had changed form my first working synthesis code
to now have dataflow fail in synthesis again. But what it boils down to is that
this code now matches more the sample HLS FFT example using front end and back
end functions to condition data prior to the FFT. So the polyphase fir operation
and phase correction have been split into seperate functions.

However, to make things a little more confusing (the whole dataflow producer
consumer relationship is a tough bug to track down) I couldn't have the apply
phase correction as a seperate function because it would keep telling me that I
had the variables read and written in more than one function. I couldn't see
what it was trying to tell me so what I did was just modify the structure of the
polyphase fir function to also do the phase correction.

Then I now have removed the decimate by 16 because like I said that was to just
produce a version that would tell me if my DMA hangs were an issue with my
pipeline and it is because it seems to be related to the decimation rate D.