Fix DMA issue
The DMA hang was caused by transferring the wrong amount of data across the DMA
at a time. For the OS PFB I knew that we needed to send D=decimation rate new
samples each time but when quickly building the PYNQ project I had been
transferring M=FFT_LENGTH worth of data. So the hangs were being caused because
the OS PFB core was not consuming all of the data the DMA was streaming to it.

Also updated the hls code input arrays to have the correct size. They were also
being synthesized with lengths M instead of D.