Improve simulation post-processing scripts
Working on debugging the output of the hardware implementation (in jupyter
notebooks) seems to show that I may be getting correct output from the OS PFB
implementation but just interpretting the data incorrectly. This has lead me to
clean up the matlab post-processing script to be more descriptive in comments
and variables to explain the corase channel plots and more importantly the fine
zoom stage and the channel discarding mechanisims.