Modify post processing scripts
To help debug workign with the hardware I returned to the simulation and post
processing code to double and triple check my understanding of the oversampled
PFB.

Previously I knew that the channels were overlapped and had to be discarded but
I wasn't applying that to the spectrum for individual channels.

Also convinced myself that when working with the output of a PFB (critically
sampled or not) it is only correct to think of the ouptut in terms of the
decimation rate. Meaning that for example a second stage channelizer and the
number of retained channels, etc. need to be interms of the decimation rate (M
or D, i.e., critically or oversampled).