Fix shift state synthesis
When testing the IP core on the hardware it was apparent that the arthimetic was
correct but that the tones would not appear in the correct bins. After working
with it for a while it would seem that the spectrum was flipped/reveresed with
something going on.

I had played with putting several different tones in different places to see
what sort of transformation needed to happen in order to get the tones to line
up in the correct place. Creating a rubic I was able to determine that in order
to get the tones to appear in the right order it looked like I would have to
apply a periodic sequence of tranformation to the indexing. It was looking like
this: FFTShift (as expected), ???, No FFTShift, ???. And I was trying to figure
out what the remaning ??? transformations would have to be.

At this point I then had the idea to stick a tone in the center of each coarse
channel such that in the second stage zoom output the channel would appear right
in the center of the spectrum. After doing this it was obvious to me what was
going on. The shift state logic between the FIR filter outputs and the IFFT was
not being applied in the hardware. The tones in bin center were period with the
number of shift states present in the oversampled ration (4/3, e.g., 4 states)
and the spectrum needed to be rotated by fourths. And so like I said, when I saw
this it was obvioius. Because the remaining ??? would be the 1/4 and 3/4
rotation to get everything to line. Which is why the FFTShift and No FFTShift
happend in the way they did. A No FFTShift would correspond to the 1/2 index
swap to re-align the channels due to the nature of the oversampled PFB.

Revisiting the HLS I noticed that the synthesis step was complaining about
non-sequential memory accesses in the filter_out variable during the shift
correeciton step. The warning also indicated that this could result in things
not being synthesized correctly. Sure enough that was happening. I modifed the
logic to now not roate the shift state array and instead reveresed the mapping
so that the filter_out memory addresses are being accessed sequentially (much
like how it was done in the FIR loops earlier by interchanging the loop indexes
to get sequential access patterns).