Separate phase rotation into seprate function
The standalone function to apply the phase rotation had not been working with
the dataflow optimization and so that step was included in the polyphase filter
method. This commit seems to have solved that issue being able to now have four
seprate methods to produce the full os pfb top function.

To get this to work required using the HLS scoped naming label to group the
functions together as to not confuse the HLS compiler as it goes through and
inserts code during the IR transformation step. After this change errors
continued to persist stating the the default mode of 'ap_auto' for the
filter_out and ifft_buffer variables were note compatible with the HLS FFT
library. To get around this additional pragmas were needed in the  the apply
phase correction method that declared the input/output arrays as ap_fifos.

The reason for this interface declaration is still a litte unclear to me because
the previous working version of the code with didn't make such declaration.
Neither did the fft_ifft HLS example application. The only idea I have is to
speculate that there are some inherited interface parameters that are inferred
and when there were three dataflow functions the chaining of function interfaces
was able to link together and infere compatible types. But the addition of
another function breaks this chain and now needs to be specified.

Also note that the polyphase filter method cannot have it's interfaces as
ap_fifo without the os pfb top level function also having it's inputs declared
as ap_fifo. There is some incompatibility here.