Accommodate fixed vs float types
The OS PFB will need to actually be a fixed point architecture for optimal
performance and resource usage. Typedefs and definitions have been rearranged and
added to allow for a fixed point implementation. The floating point
implementation has been left in for now but may be removed if it becomes to
cumbersome to continue and keep around.