Understand and implement added latency calculation
Initially I noticed that the symbolic timing print out didn't match my hand
drawn version for M=4, D=3, P=3. The difference was that in my hand drawn timing
diagram the same samples (the M-D samples, e.g., x-9, x-6, x-3) were always the
same ones repeated in the loop back delay buffer. However, in my symbolic
implementation here the samples that would be repeated in the loop back would
drift over time.

So I had set out to re-implement the logic to step each PE. However, to my
surprise after re-implementing the logic I get the same result. Instead I knew I
really needed to understand why the latency would increase as the number of PEs
increased.

After implementing my symbolic timing print out matching my hand drawn case and
before noticing the above I had could also see the phenomena occuring and knew
where I could see the extra latency showing up. It would be seen as (what was
previously) an unexpected added delay from carrying the last sample of the
previous PE into the next. I had then also noticed the previous inconsistencey
and decided to look into that one first. But since it was wrong I then came back
to this one.

I knew it was related to the modulo relationship of the resampling operation
but to what extent it included the full 2*M + M-D delay between each stage or
not I was unsure. But examining the pattern shown in my written table of
experimentally derived values it matched the same pattern seen as when you would
deliver an input sample to the polyphase fir as explained in f.j. harris or m.d.
rice.

This therefore led to the completion of the latencyComp method (not really
adequetly named since it just gets the added delay or compensation amount). What
is going on here is that the divmod method is used to determine when the modulo
operator overflows. When it does the iterator value increases otherwise when the
modulo operator doesn't overflow a delay needes to be added.

The other intersting thing is that the initial divmod value is set to M-1 and
the process starts from there. I am still working on explaining that...