Continuing to correct latency calculation
Need to clarify the comments by either cleaning them up, removing or adding more
about what has happened here. But briefly, these changes move from the init
value of x1 to x0. This process required changing the starting value of the PE
taps and modtimer counting True and False input flag valid signals.

Also required was modifying the starting values for the latency compensation
fucntion. It was also observed in this process that the startbranch of the sink
should be calculted using the later time instead of the earlier time as to
prevent having values that were not input values (e.g., having x-1 after the
first time we see x0).