Finish staging merge of the sfg/symbolic work
The SFG and symbolic OS PFB are being merged into the main OS PFB repo that I
had been previoulsy worked in. This primiarly just brings together everything
that has been done to work on the OS PFB for ALPACA.

The SFG and symbolic work took over after HLS didn't seem very viable
anymore. The idea is then to use the concepts of systolic array design to create
an SFG (which was done) prove its working symboliclly (seems to be working) and
then build out the HDL code from here.

There is still work to be done to show the rotation and can be incorporated into
the polyphase FIR structure. This is currently where the progress is at.