Comment and explain much of the OS PFB and current progress
This comment gathers a lot of the comments that were placed inside the main body
of the code and places them in the classes or methods where it is relevant.
Primiarly the biggest focus is to document the idea of the init value as it is
in limbo and undetermined if it will be present in future versions. But the idea
was to at least document it here in case wanted to go back and learn a little
more about what was going on at the time. In addition is the documenting of the
latency calculation and how those algorithms work based on the assumptions made
about the init value and its interaction with the sink class.