Initial commit on ping pong buffer
After the polyphase FIR structure a phase rotation buffer is used prior to the
FFT to correct for the phase error introduced by decimating at a rate other than
the critical rate.eq

This phase rotation will be implemented as a ping pong buffer going back and
forth between setting and fetching from them. This is because the FFT requires
that samples be sent to the FFT in natural order and therefore requires a
latency to wait the worst case number of cycles, the transform size (M) to
receive all the samples. However, the difference is that with the phase rotation
the number of samples required to wait changes and instead of reading in and out
of the ping pong buffers in the same manner each M cycles a different shift is
applied.

The underlying data structure for this buffer is a stack and to implement the
phase rotation the top and bottom both move by the shift amount after being
filled and the top is read around until reaching the bottom again. When moving
to hdl a state machines will of course be the best way to proceed but it may be
a good idea to also use the stack idea as to be able to have an arbitrary
oversample ratio if possible.

The state machine right now seems fairly straight forward but there are a few
nuances (like always starting back at zero and the access index after a shift is
always relative to zero) I briefly looked into ways not have to move the pointer
back to zero each time (start filling form where the address pointers end up)
but it did not seem to work and I decided that it really should work with
minimal latency but potentially could make it slightly a little more
complicated.