Working pingpong buffer for phase compensation
There is still work to be done to refactor the code but this commit shows the
working structure for how the pingpong buffer operates by monitoring the full
flag and marking that buffer to be pulled from. It also shows how the stack
moves the top and bottom according to the shift state to accomplish the shift
behavior.

The quesiton is how should this be incorporated into the larger ospfb. Because
right now the class name is pingpong buffer but the logic is still in the main.
If we move this into the ppbuf class then it really isn't just a ppbuf class it
is the full phase compensation. And so do we just call it phaseComp or something
like that? Or is there continued refactoring I can do to seperate out the
functionality.