pingpong buffer class implementing phase compensation operation
This commit takes the previous commit with the logic and operation of the
control and phase compensation from the main function and moves it into the ping
pong buffer class.

The current question is if it should stay in this form for verification and then
taking to HDL or should I seperate the ppbuf and phase compensation behavior out
even more? The concern being how easy it is to replicate the concurrent hardware
control based on how many software abstractions I include.

In this current form and in my mind the phase compensation state machine when
written in HDL would look much like this. At least based on other state machines
I have written and how the combinational logic woudl perform to produce the
outputs.