Seperate the phase compensation control step from the ping pong buffer
This abstracts the phase compensation step out into another class using a
generic ping pong buffer.

I am stil unsure if this is the exact architecture to mimic when coding the hdl
because I don't know if it will bite me with any software assumptions instead of
requiring concurrent hardware operations.

Note also that the ping pong buffer while is "generic" it isn't completely
implemeted out with a write and read method that would make it fully
operational as a multi-use ping pong buffer. It really is only coded for this
one use. (Again an argument to keep them combined) But it isn't too far off from
a complete generic implementation