Allow containers to operate on different data types
Implements general data types for the stack, ping pong and phase compensation
operations.

Up until now the simulatios have all been running with symbolic string
representations of the output data products. This commit implements the idea
from the beginning to allow the containers to hold whatever type (string or
numeric) so that we can then take the symbolic data outputs and show that this
implementation works as expected.

The source and types implementation are approximate at this point. In other
words, it gets the job done but a cleaner and better implementation could be
considered.