Some phase comp edits while trying to finish pfb
Add valid buffer and data buffer as part of the phase compensation processing.
Not positive at this point yet how important it is to the processing but it was
included to match and be consistent with the polyphase fir MAC PEs.

The idea being that these buffers would pass control singals. Recently, it seems
to make sense that this is still the case but the logic to generate and use the
right signal at the right time is still a little in question.

Same with the data buffer. After the polyphase fir step it seems that we woudl
be OK dropping the samples (except for raw voltage modes?) and could just leave
that dangling? But now as I am typing this -- I don't think we could use it for
raw voltage because the polyphase fir resamples the data causing duplicates and
so we would have to remove those. But as I say, may be worth thinking about. But
the data shouldn't be needed for pfb processing after the polyphase fir, that is
certain.

Additionally, started to add the history logger similar to the polyphase fir.  I
had started coding that to help debug the problem with lining up the phase
rotation and polyphase fir processing. But, was able to solve it before it was
completely ready. There were some issues that need to be re-evaluated if wanting
to put that printing logger in. The logger either needs to instead be added to
the ppbuf or stack class instead of phase comp or better thought about how to
concatenate both stacks form the ppbuf and print them out in the cycle-by-cycle
format similar to my hand drawn work.