Working OS PFB architecture
This commit is messy, but it works.

The enable signal to start the phase compensation rotation needs to be better
mapped to how it should function in hardware instead of just computing the
delay. While it seems like a state machine could just count when it should start
firing it would be better to utilize the valid buffer. Preliminary assessment
seems to indicate that should work just fine (and be more robust).

The phase compensation processing element is added at the end of the MAC
polyphase FIR PEs. The output of the last one is handed to the phase
compensation step and is activated after (P-1)*M + 2 cycles. This delay is the
delay of the sum line of the ospfb (P*M+1) minus the delays of the "wind up" to
the phase compensation step anticipating the first M samples to load it before
reading samples out. Then the additional one is to move of the zero-th sample
(e.g., P*M+1 - M + 1= (P-1)*M+2.

Recognizing the above was one of the two keys to getting the phase comp to
integrate correctly:

First was the sink was modified to generate ospfb sums in natural (traditional
parallel) or reverse (processing) order. In addition the notation I have
suggested to describe the ospfb operation was included to show all processing
steps y_m[nD] = y'(m,s)[nD], etc. Why this was relevant is because the Xilinx
FFT expects samples to come in natural order and so the output of the phase
compensation needed to come in natural order and the sink needed to generate
samples in that order (as opposed to the reverse processing order it had been
doing).

Second was a minor switch in thinking about the delay through the filter. Since
the beginning I have been worked up about knowing exactly the latency through
the filter (which I still think is valuable and a must know) but part of me is
thinking I have been thinking about it a little wrong (although, again, it is
still more convincing to be able to know the exact filter state at any time and
know when a sample goes in where it will be -- i.e., my study and concern with
knowing when the initial zero sample will appear on the output in the last tap).

And what I mean by all this is that I knew I had the polyphase fir portion
correct. I had spent the time to get it right and know when outputs would come
out that I was expecting (zero sample, last tap). So I knew that as soon as data
was coming out valid polyphase fir sums would be present and that the phase
compensation would just need to work.

So I first tried just letting the phase compensation step do its thing and free
run from the beginning. That didn't work, and it made sense because there was a
specific latency through the filter and the phase rotation states need to be on
the right samples at the right time.

So thinking about it, as I said, when polyphase fir sum outputs were produced
and valid they would just flow continually and that the phase compensation would
just need to be operating at the right state as soon as it saw data.

With this in mind it made sense then that the phase compensation step needed to
start after some delay and that it would have to be independent of the delay
term. Which this is what I saw. I first noticed with M=8, D=6, P=2 and M=8, D=5,
P=2. The delay had to be the same. It was then the idea came back to me to think
again about fifo delay on the sum term and that there is M delays per PE. This
line of thought is what I thought the delay would be at the beginning but
instead latencyComputation was in terms of including the 2*M and M-D delay on
the data lines. But (again) I knew the sum terms are what needed to be rotated
and those were always right. I also knew that the first rotation wasn't really a
rotation.

This is when I realized that delay would be M*P+1 minus anticipating the phase
compensation filter. With an additional one to move us off the zero sample. It
was here again I also realized that P*M++ is L=PM (the length of the filter) +1
and is a normal dsp thing to wait (like in how working with filtered and
convolution data). M*P+1 + 1 is also when the initial (x0) sample would come
out.

(note that the remaining M-1) port have '-' in the first term because the very
next thing we do is we insert a bunch of null "*" data.

So long story short. I still want to know what the overall latency is. But what
we actually care about isn't far off form what we know already about being the
length of the filter+1.

The code to verify just steps through and checks the sums with the sink. Instead
of knowing exactly when the rest of the data will be valid (like I did with the
polyphase fir) I instead now just go to P*M+1+1 and know that this is when the
sample I first put in is now on the output and just check as many taps of the
polyphase output as are valid with the sink until all taps have symbolic data.

One thought is to initialize the filter state in the symbolic form... but that
may be too much work.

From here I am looking forward to templating out these data containers to hold
floats like the phase compensation.