Second stage spectrum now seems to be right
Last commit thought that the remaining spur was due to not capturing the outputs
from the polyphase FIR/phase compensation step in the right sequence into the
IFFT.

Well after looking at it more it turns out the second spur was similar to what I
had seen previously when coding the OS PFB in HLS with the phase compensation
states. What it was is that the phase states was shifting in the order
opposite of what it should be. For example, with M=32, D=24 and M/D=4/3 the
code to generate the shift states is

numstates = 4
states = [n*D % M for n in range(0, numstates)]

resulting in states = [0, 24, 16, 8] and the code was progressing applying
states 0, 24, 16, 8, 0, 24, etc. However, to remove the spur the correct
sequence is 0, 8, 16, 24, 0, 8, 16, etc. I modified the states generation to
instead now be

states = [-(n*D) % M for n in range(0, numstates)]

It is important that I go back and make sure I understand why it is this way.