Initial implementation of golden model
The PFB is an efficient implementation of a bank of decimated LPF. The OSPFB is
the same implementation where the decimation rate is not equal to the number of
branches in the filter bank. When the decimation rate equals the number of
branches in the bank this is the critically sampled filter bank.

Therefore, to some degree the bank of decimated LPF is really the gold standard
in terms of validating the results. This commit includes a matlab and python
implementation of the inefficient bank of decimated LPF for arbitrary decimation
rate (critically or oversampled) to integrated as to verify the implementation
of the OSPFB.