Everything seems to be working!
The golden model is now used to compare the output of the OS PFB hardware
simulation. The simulation runs collecting multiple frames of output then
creating an averaged periogagram off the outputs of the golden model and the OS
PFB. There is a slight discrepencey between the two but we also saw this in our
matlab simulations between our intial coding of the OSPFB and a bank of
decimated low pass filters.

The golden model continues to accumulate input samples until enough samples from the
OSPFB have been collected to compute a fine spectrum. Then the golden model runs
thorugh a block of samples computing the equivalent spectrum for that interval.
The intervals seems to be lining up fine which seems to validate that I
understand the timing and the latency.

From here will continue to clean up the code and verify further. Have yet to run
at the full ALPACA spec.

I am not worried about it now but it looks like there was some interesting
processing artifacts in the golden model when running M=32 D=27. The OSPFB
output looked fine it was the golden model that had some interesting spurs going
on.