simple phase compensation buffer and test bench
For the HDL implementation I am starting with the phase compensation buffer to
come back up to speed on System Verilog and verification test bench programming.

So the test bench shows that functionally the simulation is correct producing
the correct shifted output for M=8, D=6 over two iterations.

much more to do here:
  * shift offset needs to be corrected
  * control signals
  * improve test bench testing
  * and so much more...