test bench updates
Updating the testbench with some more helpful methods and checking as to not
have to rely on the GUI and waveforms. Slower startup but shoudl be more helpful
in the long run.

The previous code wasn't really working... got to the right answer for some
cases by accident. Was having some System Verilog "gotcha" problems with the
loosely typed arithmetic.