Actual working phase compensation buffer
Finally able to get something that passes the beginning of my initial simple
testbench sending a ramp through the ping pong buffer. The loosely typed system
verilog assignment "gotchas" had me getting unexpected simulation errors.
Specfically assigning vectors of mismatching widths. The problem was I was
rolling over to an unexpected word width.

The immediate solution was to have a temporary wire that matched read address
length I was expecting.

While this works this simple test case was more a thought experiment/excerise
for me to get back into the swing of things. For one, this simple example I am
using the same behavioral RAM variable and partitioning the address space into
two. This partition takes the A and B regions of the ping pong buffer. But may
be impractical for a BRAM.

However, this does synthesize in Vivado. In the simple example I get a
distributed RAM and a specific sythesis report saying that using a BRAM would be
inefficient. I am curious now what synthesis will look like were I to make the
memory deeper. Because, we know that BRAMs can be cascaded together and the tool
may be able to take this behavioral description (since I do not instantiate
using the BRAM primitive) and cascade the BRAMs correctly.

But, I am new enough at this that I still don't have enough of an intuition or
knowledge of what the tool does to create something with that big an address
space. Then subsequently if this prododuces intermediate logic that could slow
us down somehow.

After this, I am also wanting to do a behavior version for inferring two BRAMs
and see how the tool handles it. If necessary I will then instantiate primitives
to see how much a difference it makes.

Further more, although this isn't immediately relevant to the ping pong buffer,
but we have a (rough) sysgen version of the OSPFB and the resource
utilization is much different than I was expecting to be mapped form an HDL
description. That is using SRL32s for the delay lines between PEs and 2 BRAMs
for the ping pong buffer.

An additional step after these above comments would be to experiment with shift
register behavioral logic to see if I can get the same SRL32s inferred without
an explicit inference then we may be in business. Otherwise I should also look
at the primitive template for the SRL32.

If the SRL32 proves too burdensome then continuing to do BRAMs may work out
because the appropriate slicing will keep the BRAMs mappin in the correct
range.