More hdl work before pausing to look at sysgen
With the beginnings of the phase compensator I started looking into the PE and
how to implement it. The sysgen code was able to show that the we (maybe) don't
have to use BRAMs for the delay lines and the SRL32s could work.

I therefore started on that and was able to produce code that inferred the
SRL32s. The synthesis reports there is some issues with it being to large of a
module and will be difficult for floorplanning. The current thinking is that to
be more efficient I need to remove the rst and en signals that result in large
fan out in the design.

However, it is also not clear if sysgen reports any errors and what that looks
like. I was just able to make sense and some decent comparisons between the two.
The testbench and some utilities for all of this in SV is in place.

However, with the sysgen in a decent place I am switching over to that to check
it out for a bit.